flow: Support extra padding length.
Although not required, padding can be optionally added until
the packet length is MTU bytes. A packet with extra padding
currently fails sanity checks.

Vulnerability: CVE-2020-35498
Fixes: fa8d9001a624 ("miniflow_extract: Properly handle small IP packets.")
Reported-by: Joakim Hindersson <joakim.hindersson@elastx.se>
Acked-by: Ilya Maximets <i.maximets@ovn.org>

+++ lib/flow.c
@@ -764,7 +764,7 @@ miniflow_extract(struct dp_packet *packet, struct miniflow *dst)
         if (OVS_UNLIKELY(tot_len > size || ip_len > tot_len)) {
             goto out;
         }
-        if (OVS_UNLIKELY(size - tot_len > UINT8_MAX)) {
+        if (OVS_UNLIKELY(size - tot_len > UINT16_MAX)) {
             goto out;
         }
         dp_packet_set_l2_pad_size(packet, size - tot_len);
@@ -808,8 +808,7 @@ miniflow_extract(struct dp_packet *packet, struct miniflow *dst)
         if (OVS_UNLIKELY(plen > size)) {
             goto out;
         }
-        /* Jumbo Payload option not supported yet. */
-        if (OVS_UNLIKELY(size - plen > UINT8_MAX)) {
+        if (OVS_UNLIKELY(size - plen > UINT16_MAX)) {
             goto out;
         }
         dp_packet_set_l2_pad_size(packet, size - plen);
